move util functions into lib
