+++ util/lib/timestamp.js
@@ -1,0 +1,5 @@
+const timestamp = () => {
+    return new Date().valueOf();
+}
+
+module.exports = { timestamp }
